Convert floating-point types from double to single precision
Replace double precision with default real and D-exponent literals
with E-exponent literals so Formal can run on hardware without f64
support, such as Apple GPU. Relax test tolerances and use coarser
2x grids for order-of-accuracy checks dominated by roundoff.

+++ src/formal/vector_2D_s.F90
@@ -114,7 +114,7 @@ contains
      
   module procedure vector_2D_divergence
 
-    double precision, dimension(self%cells_(x_dir)+2, self%cells_(y_dir)+2) :: div_x_term, div_y_term
+    real, dimension(self%cells_(x_dir)+2, self%cells_(y_dir)+2) :: div_x_term, div_y_term
 
     call_julienne_assert(self%consistent())
